+++ src/microservices/Microservices.IsIdentifiable/Service/IsIdentifiableQueueConsumer.cs
@@ -32,6 +32,14 @@ namespace Microservices.IsIdentifiable.Service
             if (!SafeDeserializeToMessage(header, basicDeliverEventArgs, out ExtractFileStatusMessage message))
                 return;
 
+            // We should only ever receive messages regarding anonymised images
+            if (message.Status != ExtractFileStatus.Anonymised)
+                throw new ApplicationException($"Received a message with anonymised status of {message.Status}");
+
+            // TODO(rkm 2020-02-04) Check that CTP doesn't output rooted paths. Also need to use the extract root
+            // var toProcess = new FileInfo(Path.Combine(message.ExtractionDirectory, message.AnonymisedFileName));
+
+            // The path is taken from the message, however maybe it should be FileSystemOptions|ExtractRoot in default.yaml
             // If the filename has a rooted path then the ExtractionDirectory is ignored by Path.Combine
             var toProcess = new FileInfo( Path.Combine(_extractionRoot, message.ExtractionDirectory, message.AnonymisedFileName) );